Clarify something in PizzaTree
